Revert "Merge branch 'Store-SegmentedControl' into Store-branch"
This reverts commit fc12f0151b0919c20707723bc5417e5c40fc415c, reversing
changes made to bf4115fcbee7ec6d160f31a71c85f28d95a586c1.

+++ Lafeef/Model/DailyReport.swift
@@ -12,11 +12,10 @@ struct DailyReport {
     var salesAmount:Float
     var backagingAmount:Float
     var advertismentAmount:Float
-    var collectedScore:Float
+    var collectedScore:Int
     var collectedMoney: Float
     var isPassed :Bool
     var isRewarded : Bool
     var reward: Int
-    var customerSatisfaction :[CustmerSatisfaction]
-    
+    var customerSatisfaction: [CustmerSatisfaction]
 }
